requirements.txt
torch
torchvision
timm
numpy
pillow
scikit-learn
tqdm
matplotlib
seaborn
scipy
pytorch-msssim
